+++ tests/process_monitor/test_comm_from_server.py
@@ -24,6 +24,7 @@ from mantarray_desktop_app.constants import StimulatorCircuitStatuses
 from mantarray_desktop_app.constants import UPDATES_NEEDED_STATE
 from mantarray_desktop_app.main_process import process_manager
 from mantarray_desktop_app.main_process import process_monitor
+from mantarray_desktop_app.main_process.process_monitor import _redact_from_queue_items
 from mantarray_desktop_app.utils.generic import redact_sensitive_info_from_path
 from pulse3D.constants import CUSTOMER_ACCOUNT_ID_UUID
 from pulse3D.constants import INITIAL_MAGNET_FINDING_PARAMS_UUID
@@ -669,25 +670,81 @@ def test_MantarrayProcessesMonitor__check_and_handle_server_to_main_queue__handl
     monitor_thread, *_ = test_monitor(test_process_manager)
     server_to_main_queue = test_process_manager.queue_container.from_flask
 
-    okc_process = test_process_manager.instrument_comm_process
     fw_process = test_process_manager.file_writer_process
     da_process = test_process_manager.data_analyzer_process
-    expected_okc_item = "item 1"
-    expected_fw_item = "item 2"
-    expected_da_item = "item 3"
+    okc_process = test_process_manager.instrument_comm_process
 
-    mocker.patch.object(okc_process, "hard_stop", autospec=True, return_value=expected_okc_item)
-    mocker.patch.object(fw_process, "hard_stop", autospec=True, return_value=expected_fw_item)
-    mocker.patch.object(da_process, "hard_stop", autospec=True, return_value=expected_da_item)
+    expected_okc_item = {
+        "board_0": {
+            "main_to_instrument_comm": [],
+            "instrument_comm_to_main": [
+                {
+                    "communication_type": "log",
+                    "log_level": 20,
+                    "message": "Remaining serial data in cache: [], in buffer: []",
+                },
+            ],
+            "instrument_comm_to_file_writer": [],
+        },
+        "fatal_error_reporter": [],
+    }
+    expected_fw_item = {
+        "board_0": {"instrument_comm_to_file_writer": [], "file_writer_to_data_analyzer": []},
+        # put same message in both queues to make sure both get redacted
+        "from_main_to_file_writer": [
+            {
+                "communication_type": "update_user_settings",
+                "content": {"user_password": "password_to_redact", "user_name": "username_to_redact"},
+            },
+        ],
+        "from_file_writer_to_main": [
+            {
+                "communication_type": "update_user_settings",
+                "content": {"user_password": "password_to_redact", "user_name": "username_to_redact"},
+            },
+        ],
+    }
+    expected_da_item = {
+        "board_0": {
+            "file_writer_to_data_analyzer": [
+                {
+                    "communication_type": "log",
+                    "log_level": 20,
+                    "message": "TestMessage",
+                },
+            ],
+        },
+        "from_main_to_data_analyzer": [
+            {
+                "communication_type": "log",
+                "log_level": 20,
+                "message": "TestMessage",
+            },
+        ],
+        "from_data_analyzer_to_main": [
+            {
+                "communication_type": "log",
+                "log_level": 20,
+                "message": "TestMessage",
+            },
+        ],
+    }
+
+    # setup hard_stop functions to return unredacted copies of the teardown message
+    mocker.patch.object(fw_process, "hard_stop", autospec=True, return_value=copy.deepcopy(expected_okc_item))
+    mocker.patch.object(da_process, "hard_stop", autospec=True, return_value=copy.deepcopy(expected_fw_item))
+    mocker.patch.object(okc_process, "hard_stop", autospec=True, return_value=copy.deepcopy(expected_da_item))
+
+    # make redactions on original teardown messages
+    _redact_from_queue_items(expected_okc_item)
+    _redact_from_queue_items(expected_fw_item)
+    _redact_from_queue_items(expected_da_item)
 
     communication = {"communication_type": "shutdown", "command": "hard_stop"}
     put_object_into_queue_and_raise_error_if_eventually_still_empty(communication, server_to_main_queue)
 
     mocker.patch.object(
-        process_manager,
-        "perf_counter",
-        autospec=True,
-        side_effect=[0, SUBPROCESS_SHUTDOWN_TIMEOUT_SECONDS],
+        process_manager, "perf_counter", autospec=True, side_effect=[0, SUBPROCESS_SHUTDOWN_TIMEOUT_SECONDS]
     )
 
     mocked_monitor_logger_info = mocker.patch.object(process_monitor.logger, "info", autospec=True)
@@ -695,9 +752,9 @@ def test_MantarrayProcessesMonitor__check_and_handle_server_to_main_queue__handl
     invoke_process_run_and_check_errors(monitor_thread)
 
     actual_log_message = mocked_monitor_logger_info.call_args[0][0]
-    assert expected_okc_item in actual_log_message
-    assert expected_fw_item in actual_log_message
-    assert expected_da_item in actual_log_message
+    assert str(expected_okc_item) in actual_log_message
+    assert str(expected_fw_item) in actual_log_message
+    assert str(expected_da_item) in actual_log_message
 
 
 def test_MantarrayProcessesMonitor__check_and_handle_server_to_main_queue__handles_shutdown_hard_stop_by__uploading_files_after_hard_stopping_and_joining_subprocesses(
@@ -778,7 +835,7 @@ def test_MantarrayProcessesMonitor__logs_messages_from_flask__and_redacts_mantar
     mocked_logger.assert_called_once_with(f"Communication from the Server: {expected_comm}")
 
 
-def test_MantarrayProcessesMonitor__logs_messages_from_flask__and_redacts_user_password(
+def test_MantarrayProcessesMonitor__logs_messages_from_flask__and_redacts_user_password_and_user_name(
     mocker, test_process_manager_creator, test_monitor
 ):
     test_process_manager = test_process_manager_creator(use_testing_queues=True)
@@ -788,7 +845,10 @@ def test_MantarrayProcessesMonitor__logs_messages_from_flask__and_redacts_user_p
 
     to_main_queue = test_process_manager.queue_container.from_flask
 
-    test_comm = {"communication_type": "update_user_settings", "content": {"user_password": "test_password"}}
+    test_comm = {
+        "communication_type": "update_user_settings",
+        "content": {"user_password": "test_password", "user_name": "test_username"},
+    }
     expected_comm = copy.deepcopy(test_comm)
 
     put_object_into_queue_and_raise_error_if_eventually_still_empty(test_comm, to_main_queue)
@@ -796,6 +856,7 @@ def test_MantarrayProcessesMonitor__logs_messages_from_flask__and_redacts_user_p
     confirm_queue_is_eventually_empty(to_main_queue)
 
     expected_comm["content"]["user_password"] = get_redacted_string(4)
+    expected_comm["content"]["user_name"] = get_redacted_string(4)
     mocked_logger.assert_called_once_with(f"Communication from the Server: {expected_comm}")